This server/client video game solution has a card battle feature. Currently the client UI does not show each player their opponents played cards or status effects. Modify the client to ...
Modify the server to incorporate the UI changes you just made.

In this server/client game when a card battle occurs the player should see his opponents cards on screen when they are played in CardBattle._opponentCardContainer. Therefor the server should send the card details to the opponent when a card is played. Implement this functionality